Add files via upload
HW4 is the lexer and Rda is the parsing class, the driver resides in HW4. Compile both into the same workspace. When running, type in a file name to get started. If you cannot find your file, the path the file should be at (or rather where you should put it) will be displayed.